change double to long double

+++ __tests__/mocks/CalculatorMock.cpp
@@ -7,10 +7,10 @@
 class CalculatorMock: public Calculator {
 
 public:
-    MOCK_METHOD(double, add, (const double &a, const double &b));
-    MOCK_METHOD(double, sub, (const double &a, const double &b));
-    MOCK_METHOD(double, mul, (const double &a, const double &b));
-    MOCK_METHOD(double, div, (const double &a, const double &b));
-    MOCK_METHOD(double, getLastResult, ());
+    MOCK_METHOD(long double, add, (const long double &a, const long double &b));
+    MOCK_METHOD(long double, sub, (const long double &a, const long double &b));
+    MOCK_METHOD(long double, mul, (const long double &a, const long double &b));
+    MOCK_METHOD(long double, div, (const long double &a, const long double &b));
+    MOCK_METHOD(long double, getLastResult, ());
     MOCK_METHOD(void, resetLastResult, ());
 };
