+++ src/Calculator.hpp
@@ -5,17 +5,17 @@
 
 class Calculator: public CalculatorIntreface {
 private:
-    std::optional<double> lastResult = std::nullopt;
-    double _saveLastResult(double result);
+    std::optional<long double> lastResult = std::nullopt;
+    long double _saveLastResult(long double result);
 public:
     Calculator() = default;
     virtual ~Calculator() = default;
 
-    double add(const double &a, const double &b) override;
-    double sub(const double &a, const double &b) override;
-    double mul(const double &a, const double &b) override;
-    double div(const double &a, const double &b) override;
+    long double add(const long double &a, const long double &b) override;
+    long double sub(const long double &a, const long double &b) override;
+    long double mul(const long double &a, const long double &b) override;
+    long double div(const long double &a, const long double &b) override;
 
-    double getLastResult() const override;
+    long double getLastResult() const override;
     void resetLastResult() override;
 };
